Simplify page DOM, rename files to proper conventions
