+++ website/src/lib/components/chart-common/Info.svelte
