+++ website/src/lib/components/chart-common/index.ts
@@ -1,0 +1,3 @@
+export { default as Info } from "./Info.svelte";
+export { default as Title } from "./Title.svelte";
+export { default as AttributeSelect } from "./AttributeSelect.svelte";
