--- website/src/lib/components/common/index.ts
@@ -1,3 +1,0 @@
-export { default as Info } from "./info.svelte";
-export { default as Title } from "./title.svelte";
-export { default as AttributeSelect } from "./attributeSelect.svelte";
