+++ website/src/lib/components/index.ts
@@ -1,10 +1,10 @@
 // place files you want to import through the `$lib/components` alias in this folder.
 
-import Title from "./title/title.svelte";
-import Verse1 from "./verse1/verse1.svelte";
+import Title from "./title/Title.svelte";
+import Verse1 from "./verse1/Verse1.svelte";
 import Verse2 from "./verse2/verse2.svelte";
-import Verse3 from "./verse3/verse3.svelte";
-import Verse4 from "./verse4/verse4.svelte";
+import Verse3 from "./verse3/Verse3.svelte";
+import Verse4 from "./verse4/Verse4.svelte";
 
 export const verses = {
     0: Title,
@@ -14,14 +14,14 @@ export const verses = {
     4: Verse4
 };
 
-export { default as Title } from "./title/title.svelte";
-export { default as Verse1 } from "./verse1/verse1.svelte";
+export { default as Title } from "./title/Title.svelte";
+export { default as Verse1 } from "./verse1/Verse1.svelte";
 export { default as Verse2 } from "./verse2/verse2.svelte";
-export { default as Verse3 } from "./verse3/verse3.svelte";
-export { default as Verse4 } from "./verse4/verse4.svelte";
-export { default as City } from "./city.svelte";
-export { default as Sunrise } from "./sunrise.svelte";
-export { default as Nav } from "./nav.svelte";
+export { default as Verse3 } from "./verse3/Verse3.svelte";
+export { default as Verse4 } from "./verse4/Verse4.svelte";
+export { default as City } from "./City.svelte";
+export { default as Sunrise } from "./Sunrise.svelte";
+export { default as Navbar } from "./Navbar.svelte";
 
 export { titleKeyframes } from "./title/keyframe";
 export { verse1Keyframes } from "./verse1/keyframe.svelte";
