+++ website/src/lib/index.ts
@@ -1,4 +1,4 @@
 // place files you want to import through the `$lib` alias in this folder.
 
-export { duration, mountNav } from "./nav-listeners";
+export { duration, mountNav } from "./navListeners";
 export { type Keyframe, keyframes, keyframe, getFramesCount, getFrame } from "./keyframe.svelte";
